valkey(1b): RESP connection - TCP/TLS + HELLO/AUTH handshake + socketpair test
The transport layer, mirroring cap/pg_conn.c exactly (raw blocking sockets
+ shared/tls_client.c for TLS; NO Keel client calls - Keel is the inbound
server loop + async pool + KlTls, not an outbound TCP client):

- valkey_conn.c gains the connection: vk_connect (getaddrinfo/socket/connect
  with a select timeout), io_send/io_recv (raw or KlTls-tunnelled, guarded by
  HL_VALKEY_NO_TLS), a bounded receive buffer + a NON-MOVING chunk arena for
  reply aggregates (so item arrays are never realloc-moved mid-parse),
  read_reply (parse one reply via cap/respwire, borrowed into the buffer +
  arena until the next command), the HELLO 3 / AUTH handshake with RESP2
  fallback on an unknown-command error, legacy AUTH, and SELECT db. rediss://
  runs hl_tls_client_handshake (chain+hostname verify) then tunnels all I/O.
  Credentials are not retained past connect; close() scrubs the recv buffer.
- respwire fix (root cause of a handshake failure): hl_resp_parse now treats a
  NULL buffer with avail==0 as NEED_MORE (a valid empty-recv state) instead of
  PARSE_ERR; avail==0 short-circuits to NEED_MORE before any deref.
- test_valkey_conn (socketpair, 7 cases): RESP3 HELLO, AUTH-in-HELLO, RESP2
  fallback, legacy AUTH, SELECT, auth-failure rejection, GET round-trip.
- Build wiring: test_valkey_dsn + fuzz_valkey_dsn now also link respwire.c
  (valkey_conn.c references the codec); test_respwire enum compares cast to
  int (warning-clean). Base still excludes valkey (0 symbols); fuzzers green.

+++ tests/hull/cap/test_respwire.c
@@ -60,72 +60,72 @@ UTEST(respwire_writer, integer_arg) {
 
 UTEST(respwire_parse, simple_string) {
     PARSE("+OK\r\n");
-    ASSERT_EQ(r, HL_RESP_OK);
-    ASSERT_EQ(v.type, HL_RESP_STR);
+    ASSERT_EQ((int)r, (int)HL_RESP_OK);
+    ASSERT_EQ((int)v.type, (int)HL_RESP_STR);
     ASSERT_TRUE(hl_resp_is_ok(&v));
     ASSERT_EQ(consumed, (size_t)5);
 }
 
 UTEST(respwire_parse, error) {
     PARSE("-ERR bad\r\n");
-    ASSERT_EQ(r, HL_RESP_OK);
-    ASSERT_EQ(v.type, HL_RESP_ERR);
+    ASSERT_EQ((int)r, (int)HL_RESP_OK);
+    ASSERT_EQ((int)v.type, (int)HL_RESP_ERR);
     ASSERT_EQ(v.str.len, (size_t)7);
     ASSERT_EQ(0, memcmp(v.str.p, "ERR bad", 7));
 }
 
 UTEST(respwire_parse, integer_neg) {
     PARSE(":-123\r\n");
-    ASSERT_EQ(r, HL_RESP_OK);
-    ASSERT_EQ(v.type, HL_RESP_INT);
+    ASSERT_EQ((int)r, (int)HL_RESP_OK);
+    ASSERT_EQ((int)v.type, (int)HL_RESP_INT);
     ASSERT_EQ(v.i, (int64_t)-123);
 }
 
 UTEST(respwire_parse, bulk_string_binary_safe) {
     PARSE("$3\r\na\0b\r\n");
-    ASSERT_EQ(r, HL_RESP_OK);
-    ASSERT_EQ(v.type, HL_RESP_STR);
+    ASSERT_EQ((int)r, (int)HL_RESP_OK);
+    ASSERT_EQ((int)v.type, (int)HL_RESP_STR);
     ASSERT_EQ(v.str.len, (size_t)3);
     ASSERT_EQ(0, memcmp(v.str.p, "a\0b", 3));
 }
 
 UTEST(respwire_parse, bulk_null_resp2) {
     PARSE("$-1\r\n");
-    ASSERT_EQ(r, HL_RESP_OK);
-    ASSERT_EQ(v.type, HL_RESP_NULL);
+    ASSERT_EQ((int)r, (int)HL_RESP_OK);
+    ASSERT_EQ((int)v.type, (int)HL_RESP_NULL);
 }
 
 UTEST(respwire_parse, null_resp3) {
     PARSE("_\r\n");
-    ASSERT_EQ(r, HL_RESP_OK);
-    ASSERT_EQ(v.type, HL_RESP_NULL);
+    ASSERT_EQ((int)r, (int)HL_RESP_OK);
+    ASSERT_EQ((int)v.type, (int)HL_RESP_NULL);
 }
 
 UTEST(respwire_parse, boolean_and_double) {
-    { PARSE("#t\r\n"); ASSERT_EQ(r, HL_RESP_OK); ASSERT_EQ(v.type, HL_RESP_BOOL); ASSERT_EQ(v.b, 1); }
-    { PARSE(",3.5\r\n"); ASSERT_EQ(r, HL_RESP_OK); ASSERT_EQ(v.type, HL_RESP_DOUBLE); ASSERT_TRUE(v.d > 3.4 && v.d < 3.6); }
+    { PARSE("#t\r\n"); ASSERT_EQ((int)r, (int)HL_RESP_OK); ASSERT_EQ((int)v.type, (int)HL_RESP_BOOL); ASSERT_EQ(v.b, 1); }
+    { PARSE(",3.5\r\n"); ASSERT_EQ((int)r, (int)HL_RESP_OK); ASSERT_EQ((int)v.type, (int)HL_RESP_DOUBLE); ASSERT_TRUE(v.d > 3.4 && v.d < 3.6); }
 }
 
 UTEST(respwire_parse, verbatim_strips_tag) {
     PARSE("=9\r\ntxt:hello\r\n");                /* 9 bytes: "txt:" + "hello" */
-    ASSERT_EQ(r, HL_RESP_OK);
-    ASSERT_EQ(v.type, HL_RESP_STR);
+    ASSERT_EQ((int)r, (int)HL_RESP_OK);
+    ASSERT_EQ((int)v.type, (int)HL_RESP_STR);
     ASSERT_EQ(v.str.len, (size_t)5);
     ASSERT_EQ(0, memcmp(v.str.p, "hello", 5));   /* tag "txt:" stripped */
 }
 
 UTEST(respwire_parse, bignum_kept_as_text) {
     PARSE("(3492890328409238509324850943850943825024385\r\n");
-    ASSERT_EQ(r, HL_RESP_OK);
-    ASSERT_EQ(v.type, HL_RESP_BIGNUM);
+    ASSERT_EQ((int)r, (int)HL_RESP_OK);
+    ASSERT_EQ((int)v.type, (int)HL_RESP_BIGNUM);
 }
 
 /* ── aggregates ───────────────────────────────────────────────────────── */
 
 UTEST(respwire_parse, array_of_bulk) {
     PARSE("*2\r\n$3\r\nfoo\r\n$3\r\nbar\r\n");
-    ASSERT_EQ(r, HL_RESP_OK);
-    ASSERT_EQ(v.type, HL_RESP_ARRAY);
+    ASSERT_EQ((int)r, (int)HL_RESP_OK);
+    ASSERT_EQ((int)v.type, (int)HL_RESP_ARRAY);
     ASSERT_EQ(v.arr.count, (size_t)2);
     ASSERT_EQ(0, memcmp(v.arr.items[0].str.p, "foo", 3));
     ASSERT_EQ(0, memcmp(v.arr.items[1].str.p, "bar", 3));
@@ -134,26 +134,26 @@ UTEST(respwire_parse, array_of_bulk) {
 UTEST(respwire_parse, nested_array_scan_reply) {
     /* SCAN reply: [cursor, [key1, key2]] */
     PARSE("*2\r\n$1\r\n0\r\n*2\r\n$2\r\nk1\r\n$2\r\nk2\r\n");
-    ASSERT_EQ(r, HL_RESP_OK);
-    ASSERT_EQ(v.type, HL_RESP_ARRAY);
+    ASSERT_EQ((int)r, (int)HL_RESP_OK);
+    ASSERT_EQ((int)v.type, (int)HL_RESP_ARRAY);
     ASSERT_EQ(v.arr.count, (size_t)2);
-    ASSERT_EQ(v.arr.items[1].type, HL_RESP_ARRAY);
+    ASSERT_EQ((int)v.arr.items[1].type, (int)HL_RESP_ARRAY);
     ASSERT_EQ(v.arr.items[1].arr.count, (size_t)2);
 }
 
 UTEST(respwire_parse, map_resp3) {
     PARSE("%1\r\n$1\r\nk\r\n:7\r\n");
-    ASSERT_EQ(r, HL_RESP_OK);
-    ASSERT_EQ(v.type, HL_RESP_MAP);
+    ASSERT_EQ((int)r, (int)HL_RESP_OK);
+    ASSERT_EQ((int)v.type, (int)HL_RESP_MAP);
     ASSERT_EQ(v.arr.count, (size_t)2);          /* 2*pairs */
-    ASSERT_EQ(v.arr.items[1].type, HL_RESP_INT);
+    ASSERT_EQ((int)v.arr.items[1].type, (int)HL_RESP_INT);
     ASSERT_EQ(v.arr.items[1].i, (int64_t)7);
 }
 
 UTEST(respwire_parse, empty_array) {
     PARSE("*0\r\n");
-    ASSERT_EQ(r, HL_RESP_OK);
-    ASSERT_EQ(v.type, HL_RESP_ARRAY);
+    ASSERT_EQ((int)r, (int)HL_RESP_OK);
+    ASSERT_EQ((int)v.type, (int)HL_RESP_ARRAY);
     ASSERT_EQ(v.arr.count, (size_t)0);
 }
 
@@ -161,33 +161,33 @@ UTEST(respwire_parse, empty_array) {
 
 UTEST(respwire_parse, need_more_incomplete_header) {
     PARSE("+OK");                               /* no CRLF yet */
-    ASSERT_EQ(r, HL_RESP_NEED_MORE);
+    ASSERT_EQ((int)r, (int)HL_RESP_NEED_MORE);
     ASSERT_EQ(consumed, (size_t)0);
 }
 
 UTEST(respwire_parse, need_more_partial_bulk_body) {
     PARSE("$5\r\nhel");                          /* body shorter than declared */
-    ASSERT_EQ(r, HL_RESP_NEED_MORE);
+    ASSERT_EQ((int)r, (int)HL_RESP_NEED_MORE);
 }
 
 UTEST(respwire_parse, need_more_partial_array) {
     PARSE("*2\r\n$3\r\nfoo\r\n");                /* only one of two elements */
-    ASSERT_EQ(r, HL_RESP_NEED_MORE);
+    ASSERT_EQ((int)r, (int)HL_RESP_NEED_MORE);
 }
 
 UTEST(respwire_parse, err_bad_bulk_length) {
     PARSE("$999999999999999999999\r\n");        /* length overflows i64 */
-    ASSERT_EQ(r, HL_RESP_PARSE_ERR);
+    ASSERT_EQ((int)r, (int)HL_RESP_PARSE_ERR);
 }
 
 UTEST(respwire_parse, err_bulk_missing_crlf) {
     PARSE("$3\r\nfooX\n");                       /* body not terminated by CRLF */
-    ASSERT_EQ(r, HL_RESP_PARSE_ERR);
+    ASSERT_EQ((int)r, (int)HL_RESP_PARSE_ERR);
 }
 
 UTEST(respwire_parse, err_unknown_type) {
     PARSE("^nope\r\n");
-    ASSERT_EQ(r, HL_RESP_PARSE_ERR);
+    ASSERT_EQ((int)r, (int)HL_RESP_PARSE_ERR);
 }
 
 UTEST(respwire_parse, err_too_deep_nesting) {
@@ -198,7 +198,7 @@ UTEST(respwire_parse, err_too_deep_nesting) {
     g_arena_off = 0;
     size_t consumed = 0; HlRespValue v;
     HlRespResult r = hl_resp_parse((const uint8_t *)deep, o, &consumed, &v, test_alloc, NULL);
-    ASSERT_EQ(r, HL_RESP_PARSE_ERR);
+    ASSERT_EQ((int)r, (int)HL_RESP_PARSE_ERR);
 }
 
 UTEST(respwire_parse, aggregate_without_arena_rejected) {
@@ -206,7 +206,7 @@ UTEST(respwire_parse, aggregate_without_arena_rejected) {
     size_t consumed = 0; HlRespValue v;
     const char *s = "*1\r\n:1\r\n";
     HlRespResult r = hl_resp_parse((const uint8_t *)s, strlen(s), &consumed, &v, NULL, NULL);
-    ASSERT_EQ(r, HL_RESP_PARSE_ERR);            /* no alloc -> reject aggregate */
+    ASSERT_EQ((int)r, (int)HL_RESP_PARSE_ERR);            /* no alloc -> reject aggregate */
 }
 
 UTEST_MAIN();
